split test cases for subclasses of ParquetInputFormat
